Move BaseModel conversion to `outlines.fsm`
