+++ outlines/fsm/json_schema.py
@@ -2,10 +2,10 @@ import inspect
 import json
 import re
 import warnings
-from typing import Callable, Optional, Tuple
+from typing import Callable, Optional, Tuple, Type, Union
 
 from jsonschema.protocols import Validator
-from pydantic import create_model
+from pydantic import BaseModel, create_model
 from referencing import Registry, Resource
 from referencing._core import Resolver
 from referencing.jsonschema import DRAFT202012
@@ -84,6 +84,39 @@ def build_regex_from_schema(schema: str, whitespace_pattern: Optional[str] = Non
 
     content = schema.contents
     return to_regex(resolver, content, whitespace_pattern)
+
+
+def convert_json_schema_to_str(json_schema: Union[dict, str, Type[BaseModel]]) -> str:
+    """Convert a JSON schema to a string.
+
+    Parameters
+    ----------
+    json_schema
+        The JSON schema.
+
+    Returns
+    -------
+    str
+        The JSON schema converted to a string.
+
+    Raises
+    ------
+    ValueError
+        If the schema is not a dictionary, a string or a Pydantic class.
+    """
+    if isinstance(json_schema, dict):
+        schema_str = json.dumps(json_schema)
+    elif isinstance(json_schema, str):
+        schema_str = json_schema
+    elif issubclass(json_schema, BaseModel):
+        schema_str = json.dumps(json_schema.model_json_schema())
+    else:
+        raise ValueError(
+            f"Cannot parse schema {json_schema}. The schema must be either "
+            + "a Pydantic class, a dictionary or a string that contains the JSON "
+            + "schema specification"
+        )
+    return schema_str
 
 
 def _get_num_items_pattern(min_items, max_items, whitespace_pattern):
